Move research page to app
