+++ frontend/stories/components/ResearchArchetypes.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta } from "@storybook/react";
-import ResearchArchetypes from "src/pages/content/ResearchArchetypes";
+import ResearchArchetypes from "src/app/[locale]/research/ResearchArchetypes";
 
 const meta: Meta<typeof ResearchArchetypes> = {
   title: "Components/Content/Research Archetypes Content",
