+++ frontend/stories/pages/research.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta } from "@storybook/react";
-import Research from "src/pages/research";
+import Research from "src/app/[locale]/research/page";
 
 const meta: Meta<typeof Research> = {
   title: "Pages/Research",
