+++ frontend/tests/components/ResearchMethodology.test.tsx
@@ -1,5 +1,5 @@
-import { render, screen } from "@testing-library/react";
-import ResearchMethodology from "src/pages/content/ResearchMethodology";
+import { render, screen } from "tests/react-utils";
+import ResearchMethodology from "src/app/[locale]/research/ResearchMethodology";
 
 describe("Research Content", () => {
   it("Renders without errors", () => {
